Added insert task into google calendar

+++ screens/ListAll.js
@@ -1,12 +1,11 @@
 
-import React from "react";
-import deleteBlueButton from "../assets/images/delete/trash-blue1.png";
-import checkButton from "../assets/images/check/check.png";
-import backButton from "../assets/images/back/back.png";
-import notificationIcon from "../assets/images/notification/notification-gradient-purple.png";
-import logoutIcon from "../assets/images/logout/003-logout.png";
-import Logout from './Logout';
-import moment from 'moment';
+import React from "react"
+import deleteBlueButton from "../assets/images/delete/trash-blue1.png"
+import checkButton from "../assets/images/check/check.png"
+import backButton from "../assets/images/back/back.png"
+import notificationIcon from "../assets/images/notification/notification-gradient-purple.png"
+import moment from 'moment'
+import ip from "../services/ipAddress.json"
 
 import {  Alert,
           FlatList, 
@@ -16,16 +15,19 @@ import {  Alert,
           Image,
           Modal,
           TouchableOpacity,
-          RefreshControl } from "react-native";
+          RefreshControl } from "react-native"
 
 export default class ListAll extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
       tasks: [],
+      tasksOfDay: [],
       taskAllInfo: {},
       modalVisible: false,
+      modalNotificationVisible: false,
       refreshing: false,
+      ip: ip.ip
     }
   }
 
@@ -36,95 +38,67 @@ export default class ListAll extends React.Component {
       height: 35
     },
     headerTintColor: '#FFF',
-    headerRight: (
-      <TouchableOpacity onPress={()=>
-        Alert.alert(
-          'Log out',
-          'Do you want to logout?',
-          [
-            {text: 'Cancel', onPress: () => {return null}},
-            {text: 'Confirm', onPress: () => {
-              ListAll._logout;
-            }},
-          ],
-          { cancelable: false }
-        )  
-        }>
-        <Image
-          style={{
-            width: 30,
-            height: 30,
-            margin: 10,
-            padding: 10
-          }}
-          source={logoutIcon}
-        />
-      </TouchableOpacity>
-    )
-  };
+  }
 
   async componentDidMount() {
     try {
-      const res = await fetch("http://192.168.0.10:5050/api/tasks", {
+      const res = await fetch(`http://${this.state.ip}:5050/api/tasks`, {
         method: 'GET',
         headers: {
           Accept: 'application/json',
           'Content-Type': 'application/json',
         },
       })
-      let tasks = await JSON.parse(res._bodyInit);
+      let tasks = await JSON.parse(res._bodyInit)
       tasks.forEach(item => {
-        let date = new Date(item.date).toUTCString();
-        date = date.replace("GMT","");
-        var importance = "";
+        let date = new Date(item.date).toUTCString()
+        date = date.replace("GMT","")
+        var importance = ""
         if (item.importance == 3) {
-          importance = "Low";
+          importance = "Low"
         } else if (item.importance == 2) {
-          importance = "Medium";
+          importance = "Medium"
         } else if (item.importance == 1) {
-          importance = "High";
+          importance = "High"
         }
-        let description = item.description.toLowerCase();
-        let uppercaseFirstLetter = description.charAt(0).toUpperCase();
+        let description = item.description.toLowerCase()
+        let uppercaseFirstLetter = description.charAt(0).toUpperCase()
         let stringWithoutFirstLetter = description.slice(1)
-        let upDescription =  uppercaseFirstLetter + stringWithoutFirstLetter;
+        let upDescription =  uppercaseFirstLetter + stringWithoutFirstLetter
         
-        item.importance = importance;
-        item.date = date;
-        item.title = item.title.toUpperCase();
-        item.description = upDescription;
-      });
-      let validTasks = await tasks.filter(item => item.check == false);
-      this.setState({tasks: validTasks});
+        item.importance = importance
+        item.date = date
+        item.title = item.title.toUpperCase()
+        item.description = upDescription
+      })
+      let validTasks = await tasks.filter(item => item.check == false)
+      this.setState({tasks: validTasks})
     } catch(err) { 
       console.log("err", err)
     }
   }
 
   async _setModalVisible(visible, id) {
-    this.setState({modalVisible: visible});
+    this.setState({modalVisible: visible})
     if (id == 0) {
       this.setState({taskAllInfo: {}})
     } else {
-      let task = this.state.tasks.filter(obj => obj._id == id);
-      let objTask = task[0];      
-      await this.setState({taskAllInfo: objTask});
+      let task = this.state.tasks.filter(obj => obj._id == id)
+      let objTask = task[0]      
+      await this.setState({taskAllInfo: objTask})
     }
-    
     return (
       <View style={styles.container}>
         <Modal
           animationType="slide"
           transparent={false}
           visible={this.state.modalVisible}
-          onRequestClose={() => {
-            Alert.alert('Modal has been closed.');
-          }}>
+          onRequestClose={() => {}}>
           <View>
             <View>
               <TouchableOpacity
                 onPress={() => {
-                  this._setModalVisible(!this.state.modalVisible);
+                  this._setModalVisible(!this.state.modalVisible)
                 }}>
               </TouchableOpacity>
             </View>
@@ -136,43 +110,43 @@ export default class ListAll extends React.Component {
 
   async _checkTask(item) {
     try {
-      let res = await fetch("http://192.168.0.10:5050/api/tasks/" + item._id, {
+      let res = await fetch(`http://${this.state.ip}:5050/api/tasks/` + item._id, {
         method: 'POST',
         body: JSON.stringify({
           check: true
         })
       })
       if(res.status !== 200) {
-        throw res.status;
+        throw res.status
       } else {
         Alert.alert('Your task has been checked')
       }
     } catch(err) {
-      console.log("err", err);
+      console.log("err", err)
       Alert.alert('A problem ocurred, please, try again later.')
     }
   }
 
   async _deleteTask(id) {
     try {
-      let res = await fetch("http://192.168.0.10:5050/api/tasks/" + id, {
+      let res = await fetch(`http://${this.state.ip}:5050/api/tasks/` + id, {
         method: 'DELETE'
-      });
+      })
       if(res.status !== 200) {
-        throw res.status;
+        throw res.status
       } else {
         Alert.alert('Your task has been deleted')
       }
     } catch(err) {
-      console.log("err", err);
+      console.log("err", err)
       Alert.alert('A problem ocurred, please, try again later.')
     }
   }
 
   _onRefresh = async () => {
-    this.setState({refreshing: true});
-    await this.componentDidMount();
-    this.setState({refreshing: false});
+    this.setState({refreshing: true})
+    await this.componentDidMount()
+    this.setState({refreshing: false})
   }
 
   _logout = () => {
@@ -181,14 +155,31 @@ export default class ListAll extends React.Component {
     )
   }
 
-  async _tasksOfDay() {
-    this.state.tasks.forEach(task => {
-      let currentDate = new Date().getDate();
-      let date = new Date(task.date);
-      if (currentDate == date) {
-        console.log(task)
-      }
-    })
+  async _tasksOfDay(visible) {
+    this.setState({modalNotificationVisible: visible})
+    let currentDate = new Date().getDate()
+    let tasksOfDay = this.state.tasks.filter(task =>  
+      currentDate == new Date(task.date).getDate())
+    await this.setState({tasksOfDay: tasksOfDay})
+    return (
+      <View style={styles.container}>
+        <Modal
+          animationType="slide"
+          transparent={false}
+          visible={this.state.modalNotificationVisible}
+          onRequestClose={() => {}}>
+          <View>
+            <View>
+              <TouchableOpacity
+                onPress={() => {
+                  this._tasksOfDay(!this.state.modalNotificationVisible)
+                }}>
+              </TouchableOpacity>
+            </View>
+          </View>
+        </Modal>
+      </View>
+    )
   }
 
   render() {
@@ -197,7 +188,7 @@ export default class ListAll extends React.Component {
         <View style={styles.containerHeader}>
           <TouchableOpacity  
             onPress={() => {
-              this._tasksOfDay();
+              this._tasksOfDay(true)
             }}>
             <Image
               style={styles.buttonHeader}
@@ -205,6 +196,48 @@ export default class ListAll extends React.Component {
             />
           </TouchableOpacity>
         </View>
+        
+        <Modal
+          animationType="slide"
+          presentationStyle="pageSheet"
+          position="bottom"
+          transparent={false}
+          visible={this.state.modalNotificationVisible}
+          onRequestClose={() => {
+            this._tasksOfDay(!this.state.modalNotificationVisible)
+          }}>
+          <View style={styles.containerTask}>
+            <FlatList
+              refreshControl={
+                <RefreshControl
+                  refreshing={this.state.refreshing}
+                  onRefresh={this._onRefresh}
+                />
+              }
+              data={this.state.tasksOfDay}
+              showsVerticalScrollIndicator={true}
+              keyExtractor={item => item._id}
+              renderItem={ ({ item }) =>
+                <View style={styles.containerTitle}>
+                  <View style={styles.containerTextNotification}>
+                    <Text style={styles.text}>{item.title}</Text>
+                    <Text style={styles.text}>{item.importance}</Text>
+                  </View>
+                </View>
+              }
+            />
+            <TouchableOpacity
+              onPress={() => {
+                this._tasksOfDay(!this.state.modalNotificationVisible)
+              }}>
+              <Image
+                style={styles.button}
+                source={backButton}
+              />
+            </TouchableOpacity>
+          </View>
+        </Modal>
+       
         <Modal
           animationType="slide"
           presentationStyle="pageSheet"
@@ -212,25 +245,23 @@ export default class ListAll extends React.Component {
           transparent={false}
           visible={this.state.modalVisible}
           onRequestClose={() => {
-            this._setModalVisible(!this.state.modalVisible, 0);
+            this._setModalVisible(!this.state.modalVisible, 0)
           }}>
           <View style={styles.containerTask}>
-            <View style={styles.modalContainer}>
-              <View style={styles.containerInfoTask}>
-                <View style={styles.text}>
-                  <Text style={styles.textTitle}>{this.state.taskAllInfo.title}</Text>
-                  <Text style={styles.textTitle}>Description:</Text> 
-                  <Text style={styles.text}>{this.state.taskAllInfo.description}</Text>
-                  <Text style={styles.textTitle}>When:</Text>
-                  <Text style={styles.text}>{this.state.taskAllInfo.date}</Text>
-                  <Text style={styles.textTitle}>Importance:</Text> 
-                  <Text style={styles.text}>{this.state.taskAllInfo.importance}</Text>
-                </View>
+            <View style={styles.containerInfoTask}>
+              <View style={styles.text}>
+                <Text style={styles.textTitle}>{this.state.taskAllInfo.title}</Text>
+                <Text style={styles.textTitle}>Description:</Text> 
+                <Text style={styles.text}>{this.state.taskAllInfo.description}</Text>
+                <Text style={styles.textTitle}>When:</Text>
+                <Text style={styles.text}>{this.state.taskAllInfo.date}</Text>
+                <Text style={styles.textTitle}>Importance:</Text> 
+                <Text style={styles.text}>{this.state.taskAllInfo.importance}</Text>
               </View>
             </View>
             <TouchableOpacity
               onPress={() => {
-                this._setModalVisible(!this.state.modalVisible, 0);
+                this._setModalVisible(!this.state.modalVisible, 0)
               }}>
               <Image
                 style={styles.button}
@@ -253,12 +284,12 @@ export default class ListAll extends React.Component {
             <View style={styles.container}>
               <TouchableOpacity
                 onPress={() => {
-                  this._setModalVisible(true, item._id);
+                  this._setModalVisible(true, item._id)
                 }}>
                 <View style={styles.containerTitle}>
                   <TouchableOpacity  
                     onPress={() => {
-                      this._checkTask(item);
+                      this._checkTask(item)
                     }}>
                     <Image
                       style={styles.button}
@@ -269,35 +300,41 @@ export default class ListAll extends React.Component {
                   <View style={styles.containerText}>
                     <Text style={styles.text}>{item.title}</Text>
                   </View>
-                  
-                  <TouchableOpacity 
-                    onPress={() => {
-                      this._deleteTask(item._id);
-                    }}>
-                    <Image
-                      style={styles.button}
-                      source={deleteBlueButton}
-                    />
+                   <TouchableOpacity onPress={()=>
+                      Alert.alert(
+                        'Delete',
+                        'Do you want to delete the task?',
+                        [
+                          {text: 'Cancel', onPress: () => {return null}},
+                          {text: 'Yes', onPress: () => {
+                            this._deleteTask(item._id)
+                          }},
+                        ],
+                        {cancelable: false}
+                      )  
+                      }>
+                      <Image
+                        style={styles.button}
+                        source={deleteBlueButton}
+                      />
                   </TouchableOpacity>
-
                 </View>
               </TouchableOpacity>
             </View>
           }
         />
       </View>
-    );
+    )
   }
 }
 
 const styles = StyleSheet.create({
   container: {
-   flex: 1,
-   paddingTop: 10
+   flex: 1
   },
   containerTitle: {
     flexDirection: 'row',
-    padding: 10,
+    margin: 10,
     alignSelf: 'auto',
     justifyContent: 'space-around',
     borderWidth: 1,
@@ -314,6 +351,26 @@ const styles = StyleSheet.create({
     alignSelf: 'flex-start',
     width: 300
   },
+  containerTextNotification: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: 300
+  },
+  containerTask: {
+    flex: 1,
+    paddingTop: 22,
+    alignItems: 'center',
+  },
+  containerModal: {
+    width: 400, 
+    alignSelf: 'center'
+  },
+  containerInfoTask: {
+    alignSelf: 'auto',
+    borderWidth: 1,
+    borderColor: '#edf4f4',
+    margin: 10,
+  },
   text: {
     fontSize: 14,
     color: 'gray',
@@ -326,22 +383,6 @@ const styles = StyleSheet.create({
     padding: 10,
     textAlign: 'justify',
     fontWeight: 'bold'
-  },
-  modalContainer: {
-    width: 400, 
-    alignSelf: 'center',
-    padding: 10,
-    margin: 10
-  },
-  containerTask: {
-    flex: 1,
-    paddingTop: 22,
-    alignItems: 'center',
-  },
-  containerInfoTask: {
-    alignSelf: 'auto',
-    borderWidth: 1,
-    borderColor: '#edf4f4'
   },
   button: {
     width: 20,
