+++ screens/ListChecked.js
@@ -1,10 +1,9 @@
 
-import React from "react";
-import deleteBlueButton from "../assets/images/delete/trash-blue1.png";
-import uncheckedButton from "../assets/images/uncheck/unchecked-gradient-purple.png";
-import backButton from "../assets/images/back/back.png";
-import logoutIcon from "../assets/images/logout/exit-gradient-purple1.png";
-import Logout from './Logout';
+import React from "react"
+import deleteBlueButton from "../assets/images/delete/trash-blue1.png"
+import uncheckedButton from "../assets/images/uncheck/unchecked-gradient-purple.png"
+import backButton from "../assets/images/back/back.png"
+import ip from "../services/ipAddress.json"
 
 import {  Alert,
           FlatList, 
@@ -14,7 +13,7 @@ import {  Alert,
           Image,
           Modal,
           TouchableOpacity,
-          RefreshControl } from "react-native";
+          RefreshControl } from "react-native"
 
 export default class ListChecked extends React.Component {
   constructor(props) {
@@ -24,6 +23,7 @@ export default class ListChecked extends React.Component {
       taskAllInfo: {},
       modalVisible: false,
       refreshing: false,
+      ip: ip.ip
     }
   }
 
@@ -34,41 +34,16 @@ export default class ListChecked extends React.Component {
       height: 35
     },
     headerTintColor: '#FFF',
-    headerRight: (
-      <TouchableOpacity onPress={()=>
-        Alert.alert(
-          'Log out',
-          'Do you want to logout?',
-          [
-            {text: 'Cancel', onPress: () => {return null}},
-            {text: 'Confirm', onPress: () => {
-              ListChecked._logout;
-            }},
-          ],
-          { cancelable: false }
-        )  
-        }>
-        <Image
-          style={{
-            width: 30,
-            height: 30,
-            margin: 10,
-            padding: 10
-          }}
-          source={logoutIcon}
-        />
-      </TouchableOpacity>
-    )
-  };
+  }
 
   async _setModalVisible(visible, id) {
-    this.setState({modalVisible: visible});
+    this.setState({modalVisible: visible})
     if (id == 0) {
       this.setState({taskAllInfo: {}})
     } else {
-      let task = this.state.tasks.filter(obj => obj._id == id);
-      let objTask = task[0];      
-      await this.setState({taskAllInfo: objTask});
+      let task = this.state.tasks.filter(obj => obj._id == id)
+      let objTask = task[0]      
+      await this.setState({taskAllInfo: objTask})
     }
     
     return (
@@ -77,14 +52,12 @@ export default class ListChecked extends React.Component {
           animationType="slide"
           transparent={false}
           visible={this.state.modalVisible}
-          onRequestClose={() => {
-            Alert.alert('Modal has been closed.');
-          }}>
+          onRequestClose={() => {}}>
           <View>
             <View>
               <TouchableOpacity
                 onPress={() => {
-                  this._setModalVisible(!this.state.modalVisible);
+                  this._setModalVisible(!this.state.modalVisible)
                 }}>
               </TouchableOpacity>
             </View>
@@ -96,43 +69,43 @@ export default class ListChecked extends React.Component {
 
   async _uncheckedTask(item) {
     try {
-      let res = await fetch("http://192.168.0.10:5050/api/tasks/" + item._id, {
+      let res = await fetch(`http://${this.state.ip}:5050/api/tasks/` + item._id, {
         method: 'POST',
         body: JSON.stringify({
           check: false
         })
       })
       if(res.status !== 200) {
-        throw res.status;
+        throw res.status
       } else {
         Alert.alert('Your task has been unchecked')
       }
     } catch(err) {
-      console.log("err", err);
+      console.log("err", err)
       Alert.alert('A problem ocurred, please, try again later.')
     }
   }
 
   async _deleteTask(id) {
     try {
-      let res = await fetch("http://192.168.0.10:5050/api/tasks/" + id, {
+      let res = await fetch(`http://${this.state.ip}:5050/api/tasks/` + id, {
         method: 'DELETE'
-      });
+      })
       if(res.status !== 200) {
-        throw res.status;
+        throw res.status
       } else {
         Alert.alert('Your task has been deleted')
       }
     } catch(err) {
-      console.log("err", err);
+      console.log("err", err)
       Alert.alert('A problem ocurred, please, try again later.')
     }
   }
 
   _onRefresh = async () => {
-    this.setState({refreshing: true});
-    await this.componentDidMount();
-    this.setState({refreshing: false});
+    this.setState({refreshing: true})
+    await this.componentDidMount()
+    this.setState({refreshing: false})
   }
 
   _logout = () => {
@@ -143,37 +116,37 @@ export default class ListChecked extends React.Component {
   
   async componentDidMount() {
     try {
-      const res = await fetch("http://192.168.0.10:5050/api/tasks", {
+      const res = await fetch(`http://${this.state.ip}:5050/api/tasks`, {
         method: 'GET',
         headers: {
           Accept: 'application/json',
           'Content-Type': 'application/json',
         },
       })
-      let tasks = await JSON.parse(res._bodyInit);
+      let tasks = await JSON.parse(res._bodyInit)
       tasks.forEach(item => {
-        let date = new Date(item.date).toUTCString();
-        date = date.replace("GMT","");
-        var importance = "";
+        let date = new Date(item.date).toUTCString()
+        date = date.replace("GMT","")
+        var importance = ""
         if (item.importance == 3) {
-          importance = "Low";
+          importance = "Low"
         } else if (item.importance == 2) {
-          importance = "Medium";
+          importance = "Medium"
         } else if (item.importance == 1) {
-          importance = "High";
+          importance = "High"
         }
-        let description = item.description.toLowerCase();
-        let uppercaseFirstLetter = description.charAt(0).toUpperCase();
+        let description = item.description.toLowerCase()
+        let uppercaseFirstLetter = description.charAt(0).toUpperCase()
         let stringWithoutFirstLetter = description.slice(1)
-        let upDescription =  uppercaseFirstLetter + stringWithoutFirstLetter;
+        let upDescription =  uppercaseFirstLetter + stringWithoutFirstLetter
         
-        item.importance = importance;
-        item.date = date;
-        item.title = item.title.toUpperCase();
-        item.description = upDescription;
-      });
-      let validTasks = await tasks.filter(item => item.check == true);
-      this.setState({tasks: validTasks});
+        item.importance = importance
+        item.date = date
+        item.title = item.title.toUpperCase()
+        item.description = upDescription
+      })
+      let validTasks = await tasks.filter(item => item.check == true)
+      this.setState({tasks: validTasks})
     } catch(err) { 
       console.log("err", err)
     }
@@ -189,25 +162,23 @@ export default class ListChecked extends React.Component {
           transparent={false}
           visible={this.state.modalVisible}
           onRequestClose={() => {
-            this._setModalVisible(!this.state.modalVisible, 0);
+            this._setModalVisible(!this.state.modalVisible, 0)
           }}>
           <View style={styles.containerTask}>
-            <View style={styles.modalContainer}>
-              <View style={styles.containerInfoTask}>
-                <View style={styles.text}>
-                  <Text style={styles.textTitle}>{this.state.taskAllInfo.title}</Text>
-                  <Text style={styles.textTitle}>Description:</Text> 
-                  <Text style={styles.text}>{this.state.taskAllInfo.description}</Text>
-                  <Text style={styles.textTitle}>When:</Text>
-                  <Text style={styles.text}>{this.state.taskAllInfo.date}</Text>
-                  <Text style={styles.textTitle}>Importance:</Text> 
-                  <Text style={styles.text}>{this.state.taskAllInfo.importance}</Text>
-                </View>
+            <View style={styles.containerInfoTask}>
+              <View style={styles.text}>
+                <Text style={styles.textTitle}>{this.state.taskAllInfo.title}</Text>
+                <Text style={styles.textTitle}>Description:</Text> 
+                <Text style={styles.text}>{this.state.taskAllInfo.description}</Text>
+                <Text style={styles.textTitle}>When:</Text>
+                <Text style={styles.text}>{this.state.taskAllInfo.date}</Text>
+                <Text style={styles.textTitle}>Importance:</Text> 
+                <Text style={styles.text}>{this.state.taskAllInfo.importance}</Text>
               </View>
             </View>
             <TouchableOpacity
               onPress={() => {
-                this._setModalVisible(!this.state.modalVisible, 0);
+                this._setModalVisible(!this.state.modalVisible, 0)
               }}>
               <Image
                 style={styles.button}
@@ -230,12 +201,12 @@ export default class ListChecked extends React.Component {
             <View style={styles.container}>
               <TouchableOpacity
                 onPress={() => {
-                  this._setModalVisible(true, item._id);
+                  this._setModalVisible(true, item._id)
                 }}>
                 <View style={styles.containerTitle}>
                   <TouchableOpacity  
                     onPress={() => {
-                      this._uncheckedTask(item);
+                      this._uncheckedTask(item)
                     }}>
                     <Image
                       style={styles.button}
@@ -247,14 +218,23 @@ export default class ListChecked extends React.Component {
                     <Text style={styles.text}>{item.title}</Text>
                   </View>
                   
-                  <TouchableOpacity 
-                    onPress={() => {
-                      this._deleteTask(item._id);
-                    }}>
-                    <Image
-                      style={styles.button}
-                      source={deleteBlueButton}
-                    />
+                   <TouchableOpacity onPress={()=>
+                      Alert.alert(
+                        'Delete',
+                        'Do you want to delete the task?',
+                        [
+                          {text: 'Cancel', onPress: () => {return null}},
+                          {text: 'Yes', onPress: () => {
+                            this._deleteTask(item._id)
+                          }},
+                        ],
+                        {cancelable: false}
+                      )  
+                      }>
+                      <Image
+                        style={styles.button}
+                        source={deleteBlueButton}
+                      />
                   </TouchableOpacity>
 
                 </View>
@@ -263,18 +243,17 @@ export default class ListChecked extends React.Component {
           }
         />
       </View>
-    );
+    )
   }
 }
 
 const styles = StyleSheet.create({
   container: {
-   flex: 1,
-   paddingTop: 10
+   flex: 1
   },
   containerTitle: {
     flexDirection: 'row',
-    padding: 10,
+    margin: 10,
     alignSelf: 'auto',
     justifyContent: 'space-around',
     borderWidth: 1,
@@ -302,7 +281,6 @@ const styles = StyleSheet.create({
     width: 400, 
     alignSelf: 'center',
     padding: 10,
-    margin: 10
   },
   containerTask: {
     flex: 1,
@@ -312,7 +290,8 @@ const styles = StyleSheet.create({
   containerInfoTask: {
     alignSelf: 'auto',
     borderWidth: 1,
-    borderColor: '#edf4f4'
+    borderColor: '#edf4f4',
+    margin: 10,
   },
   button: {
     width: 20,
